Pulled it out.  Works with xml in /deploy, but not as a bundle and feature :(
