+++ islandora-connector-derivative/build.gradle
@@ -1,23 +1,17 @@
 apply plugin: 'osgi'
 
-description = 'Islandora CLAW Fcrepo Indexer'
+description = 'Islandora CLAW Derivative Connector'
 
 dependencies {
-    compile group: 'org.apache.camel', name: 'camel-core', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-blueprint', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-http4', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-jackson', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-jsonpath', version: camelVersion
-    compile group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
     compile group: 'commons-io', name: 'commons-io', version: commonsIoVersion 
+    compile group: 'org.apache.camel', name: 'camel-blueprint', version: camelVersion
+    compile group: 'org.apache.camel', name: 'camel-core', version: camelVersion
+    compile group: 'org.apache.camel', name: 'camel-jackson', version: camelVersion
+    compile group: 'org.apache.camel', name: 'camel-http4', version: camelVersion
+    compile group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
     testCompile group: 'org.apache.camel', name: 'camel-test-blueprint', version: camelVersion
     testCompile group: 'org.apache.servicemix.bundles', name: 'org.apache.servicemix.bundles.xerces', version: '2.11.0_1'
     testCompile group: 'org.ow2.asm', name: 'asm-commons', version: '5.0.3'
-    testCompile group: 'com.googlecode.junit-toolbox', name: 'junit-toolbox', version: '2.3'
-}
-
-test {
-    maxParallelForks = Runtime.getRuntime().availableProcessors()
 }
 
 jar {
@@ -28,16 +22,8 @@ jar {
       license project.license
 
       instruction 'Import-Package', 'org.apache.camel.component.http4,' +
-                            "org.apache.camel.jsonpath," +
                             "org.apache.camel;version=\"${camelVersion}\"," +
                             defaultOsgiImports
-      instruction 'Export-Package', 'ca.islandora.alpaca.indexing.fcrepo'
-    }
-}
-
-artifacts {
-    archives (file('build/cfg/main/ca.islandora.alpaca.indexing.fcrepo.cfg')) {
-        classifier 'configuration'
-        type 'cfg'
+      instruction 'Export-Package', 'ca.islandora.alpaca.connector.derivative'
     }
 }
