+++ islandora-connector-derivative/src/test/resources/log4j.properties
