+++ islandora-connector-houdini/build.gradle
@@ -2,18 +2,6 @@ apply plugin: 'osgi'
 
 description = 'Islandora CLAW Houdini Connector'
 
-dependencies {
-    compile group: 'commons-io', name: 'commons-io', version: commonsIoVersion 
-    compile group: 'org.apache.camel', name: 'camel-blueprint', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-core', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-jackson', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-http4', version: camelVersion
-    compile group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
-    testCompile group: 'org.apache.camel', name: 'camel-test-blueprint', version: camelVersion
-    testCompile group: 'org.apache.servicemix.bundles', name: 'org.apache.servicemix.bundles.xerces', version: '2.11.0_1'
-    testCompile group: 'org.ow2.asm', name: 'asm-commons', version: '5.0.3'
-}
-
 jar {
     manifest {
       description project.description
@@ -22,9 +10,8 @@ jar {
       license project.license
 
       instruction 'Import-Package', 'org.apache.camel.component.http4,' +
-                            "org.apache.camel;version=\"${camelVersion}\"," +
+                            "ca.islandora.alpaca.connector.derivative," +
                             defaultOsgiImports
-      instruction 'Export-Package', 'ca.islandora.connector.houdini'
     }
 }
 
@@ -34,4 +21,3 @@ artifacts {
         type 'cfg'
     }
 }
-
