+++ islandora-connector-houdini/src/main/cfg/ca.islandora.alpaca.connector.houdini.cfg
@@ -4,4 +4,4 @@ error.maxRedeliveries=10
 # The camel URI for the incoming message stream.
 in.stream=activemq:queue:islandora-connector-houdini
 
-houdini.convert.url = http://localhost:8000/houdini/convert
+derivative.service.url = http://localhost:8000/houdini/convert
