+++ islandora-indexing-triplestore/build.gradle
@@ -25,7 +25,7 @@ jar {
                             "org.apache.camel;version=\"${camelVersion}\"," +
                             "org.fcrepo.camel.processor," +
                             defaultOsgiImports
-      instruction 'Export-Package', 'ca.islandora.indexing.triplestore'
+      instruction 'Export-Package', 'ca.islandora.alpaca.indexing.triplestore'
     }
 }
