+++ settings.gradle
@@ -1,11 +1,13 @@
 include ':islandora-karaf'
 include ':islandora-indexing-triplestore'
 include ':islandora-indexing-fcrepo'
+include ':islandora-connector-derivative'
 include ':islandora-connector-houdini'
 include ':islandora-http-client'
 
 project(':islandora-karaf').projectDir = "$rootDir/karaf" as File
 project(':islandora-indexing-triplestore').projectDir = "$rootDir/islandora-indexing-triplestore" as File
 project(':islandora-indexing-fcrepo').projectDir = "$rootDir/islandora-indexing-fcrepo" as File
+project(':islandora-connector-derivative').projectDir = "$rootDir/islandora-connector-derivative" as File
 project(':islandora-connector-houdini').projectDir = "$rootDir/islandora-connector-houdini" as File
 project(':islandora-http-client').projectDir = "$rootDir/islandora-http-client" as File
